[#492] AppListBuilderTest should pass integration tests.
